Update suppliers feature and design
Moved add supplier modal to another page as well as the update supplier
feature. The modals were tricky and problematic in handling errors.
In-line editing will be implemented soon, but updating via another page
will be used for now.

Errors are already added, too for each form (add and update)